histogram: port to CUDA 12; fix NPP buffer-size arg and iteration timing
Removes the unused texture-reference machinery (texture<>/cudaArray/bind)
that no longer compiles under CUDA 12, fixes the nppiHistogramEven
buffer-size argument type, and fixes TestHistogram's timing loop (it
overwrote ms every iteration, so cIterations inflated throughput ~200x).

+++ histogram/histogramNPP.cuh
@@ -59,7 +59,7 @@ GPUhistogramNPP(
     cuda(EventCreate( &stop, 0 ) );
 
     // create device scratch buffer for nppiHistogram
-    int nDeviceBufferSize;
+    size_t nDeviceBufferSize;
     nppiHistogramEvenGetBufferSize_8u_C1R(oSizeROI, levelCount ,&nDeviceBufferSize);
     cuda(Malloc((void **)&pDeviceBuffer, nDeviceBufferSize) );
